Extract markdown to file
