+++ Document/Sources/Document/Models/A11yDescription+VoiceOver.swift
@@ -1,35 +1,6 @@
 import Foundation
 import AppKit
 
-@available(macOS 12, *)
-extension NSMutableAttributedString {
-    func append(markdown: String) {
-        if markdown.isEmpty {
-            return // Do nothing, otherwise it will crash
-        }
-        append(try! NSAttributedString(markdown: markdown))
-    }
-    
-    static func += (lhs: NSMutableAttributedString, rhs: String) {
-        lhs.append(markdown: rhs)
-    }
-    
-    var isEmpty: Bool {
-        length == 0
-    }
-}
-
-// MARK: Markdown
-extension String {
-    var bold: String {
-        "**\(self)**"
-    }
-    
-    var italic: String {
-        "*\(self)*"
-    }
-}
-
 extension A11yDescription {
     
     @available(macOS 12, *)
